requirements.txt
discord.py
GitPython
Pillow